Move e2e tests out of solution and fix tests
