+++ DAPM.Backend.Test.EndToEnd/DAPM.ClientApi.Test.EndToEnd/Models/GetOrganizationsTicketResult.cs
