+++ DAPM.Backend.Test.EndToEnd/DAPM.ClientApi.Test.EndToEnd/Models/Resource.cs
